+++ src/components/CreateNewTodo.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 
 import { useTodoStore } from "@/store/todoStore";
 import { createTodo, postFile, updateTodo } from "@/api/todoAPI";
-import useModal from "@/hook/useModal";
+import useModal from "@/hooks/useModal";
 import { getGoals } from "@/api/goalAPI";
 import { GoalType, InitialTodoType, TodoType } from "@/type";
